+++ src/ufm/fine_tuning.py
@@ -1,15 +1,14 @@
 '''
 Scripts for fine-tuning on the harmful datasets
 '''
-from typing import TYPE_CHECKING
-from datasets import load_dataset, Dataset
-from transformers import AutoTokenizer, AutoModelForCausalLM
-from transformers import TrainingArguments, Trainer
 from logging import Logger
-import wandb
+from typing import TYPE_CHECKING
+
 from data import get_hf_data
-from models import HuggingFaceModel  # HF model is a wrapper with model AND tokenizer
 from omegaconf.errors import ValidationError
+from transformers import Trainer, TrainingArguments
+
+from models import HuggingFaceModel  # HF model is a wrapper with model AND tokenizer
 
 if TYPE_CHECKING:
     from omegaconf import DictConfig
@@ -64,7 +63,7 @@ def run_fine_tune(
 
     # Load dataset
     logger.info(f"Loading dataset {config['dataset']} ...")
-    dataset = get_hf_data(dataset_identifier) #TODO batch size config etc
+    dataset = get_hf_data(dataset_identifier)  # TODO batch size config etc
 
     # assert train splits exist
     assert 'train' in dataset
